Renamed both services to same name, but moved away from /services path.
This way window-mock will automatically override window service when test-support.js is loaded, as it has the same module name.
