+++ addon/-private/services/window.js
