--- app/services/window-browser.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-window-mock/services/window-browser';
